Restructured package hierarchy
